# src/main/java/com/vass/shop/infrastructure/adapters/inbound/rest/api/UsersApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.23.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.vass.shop.infrastructure.adapters.inbound.rest.api;

import com.vass.shop.infrastructure.adapters.inbound.rest.dto.ErrorModel;
import org.springframework.lang.Nullable;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.ProductResponse;
import java.util.UUID;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.UserRequest;
import com.vass.shop.infrastructure.adapters.inbound.rest.dto.UserResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-08T22:59:31.432094100+03:00[Europe/Bucharest]", comments = "Generator version: 7.23.0")
@Validated
@Tag(name = "users", description = "the users API")
public interface UsersApi {

    String PATH_CREATE_USER = "/users";
    /**
     * POST /users : Create a user
     *
     * @param userRequest  (required)
     * @return User created successfully (status code 201)
     *         or Invalid request payloads or schema errors. (status code 400)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "createUser",
        summary = "Create a user",
        responses = {
            @ApiResponse(responseCode = "201", description = "User created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payloads or schema errors.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = UsersApi.PATH_CREATE_USER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<UserResponse> createUser(
        @Parameter(name = "UserRequest", description = "", required = true) @Valid @RequestBody UserRequest userRequest
    );


    String PATH_DELETE_USER = "/users/{userId}";
    /**
     * DELETE /users/{userId} : Delete a user
     *
     * @param userId  (required)
     * @return User deleted successfully (status code 204)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Delete a user",
        responses = {
            @ApiResponse(responseCode = "204", description = "User deleted successfully"),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = UsersApi.PATH_DELETE_USER,
        produces = { "application/json" }
    )
    ResponseEntity<Void> deleteUser(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    );


    String PATH_GET_TOP_EXPENSIVE_PURCHASES = "/users/{userId}/history/top-expensive";
    /**
     * GET /users/{userId}/history/top-expensive : Fetch top 4 most expensive single products purchased historically by user
     *
     * @param userId  (required)
     * @return Array of top 4 expensive items purchased (status code 200)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "getTopExpensivePurchases",
        summary = "Fetch top 4 most expensive single products purchased historically by user",
        responses = {
            @ApiResponse(responseCode = "200", description = "Array of top 4 expensive items purchased", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProductResponse.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_GET_TOP_EXPENSIVE_PURCHASES,
        produces = { "application/json" }
    )
    ResponseEntity<List<ProductResponse>> getTopExpensivePurchases(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    );


    String PATH_GET_USER_BY_ID = "/users/{userId}";
    /**
     * GET /users/{userId} : Get user by ID
     *
     * @param userId  (required)
     * @return User details found (status code 200)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "getUserById",
        summary = "Get user by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "User details found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_GET_USER_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<UserResponse> getUserById(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    );


    String PATH_LIST_USERS = "/users";
    /**
     * GET /users : List all users
     *
     * @param limit How many users to return at one time (optional, default to 10)
     * @return A list of users (status code 200)
     *         or Invalid request payloads or schema errors. (status code 400)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "listUsers",
        summary = "List all users",
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of users", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = UserResponse.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payloads or schema errors.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_LIST_USERS,
        produces = { "application/json" }
    )
    ResponseEntity<List<UserResponse>> listUsers(
        @Parameter(name = "limit", description = "How many users to return at one time", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "10") Integer limit
    );


    String PATH_UPDATE_USER = "/users/{userId}";
    /**
     * PUT /users/{userId} : Update user details
     *
     * @param userId  (required)
     * @param userRequest  (required)
     * @return User updated successfully (status code 200)
     *         or Invalid request payloads or schema errors. (status code 400)
     *         or Authorization header missing or validation failed. (status code 401)
     *         or Requested domain entity could not be found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Update user details",
        responses = {
            @ApiResponse(responseCode = "200", description = "User updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payloads or schema errors.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authorization header missing or validation failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Requested domain entity could not be found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorModel.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = UsersApi.PATH_UPDATE_USER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<UserResponse> updateUser(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId,
        @Parameter(name = "UserRequest", description = "", required = true) @Valid @RequestBody UserRequest userRequest
    );

}